CSbackend/storage/serializers.py
# storage/serializers.py

from rest_framework import serializers
from .models import File

class FileSerializer(serializers.ModelSerializer):
    class Meta:
        model = File
        fields = ['id', 'original_name', 'size', 'upload_date', 'last_download_date', 'comment', 'special_link']
