CSbackend/requirements.txt
# requirements.txt

gunicorn>=20.1.0
dj-database-url>=0.5.0
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
